feat: 판매자 상품 수정 기능을  추가합니다

+++ src/components/console/console-main/console-main.tsx
@@ -1,33 +1,43 @@
-import { FirebaseError } from 'firebase/app';
-
 import { Li, Ul } from '@components/common/list';
 import Wrapper from '@components/common/wrapper';
 import { Button } from '@components/ui/button';
 
-import type { Product } from '@src/lib/firebase/types';
 import { useDeleteProductsMutation } from '@hooks/use-console-products-query';
 
+import { useModalContext } from '@providers/modal';
+
+import type { Product } from '@src/lib/firebase/types';
+import ConsoleUpdateProduct from '../console-update-product/console-update-product';
+
 interface Props {
-  products?: Product[];
+  products: Product[];
 }
 
 export default function ConsoleMain({ products }: Props) {
   const { mutate: deleteProduct } = useDeleteProductsMutation();
+  const { openModal } = useModalContext();
+
+  const onClickEditButton = (productId: string) => {
+    const product = products.find((p) => p.id === productId);
+
+    if (!product) {
+      throw new Error('product not found');
+    }
+
+    openModal(<ConsoleUpdateProduct product={product} />);
+  };
 
   const onClickDeleteButton = async (productId: string, sellerId: string) => {
-    try {
+    const confirm = window.confirm('정말로 삭제하겠습니까?');
+    if (confirm) {
       deleteProduct({ sellerId, productId });
-    } catch (error) {
-      if (error instanceof FirebaseError) {
-        console.error(error);
-      }
     }
   };
 
   return (
-    <Ul className="flex-col items-start gap-4 px-4 pb-4">
-      <h1 className="text-3xl font-bold">전체 상품 조회</h1>
-      {products?.map(({ createdAt, productName, productDescription, thumbnailUrl, id, sellerId }) => {
+    <Ul className="flex-col items-start space-y-4">
+      <h1 className="px-4 text-2xl font-bold">전체 상품 조회</h1>
+      {products?.map(({ createdAt, productName, productDescription, thumbnail, id, sellerId }) => {
         const creaatedAt = createdAt.toDate();
         const date = new Intl.DateTimeFormat('ko-KR', {
           dateStyle: 'full',
@@ -36,25 +46,28 @@ export default function ConsoleMain({ products }: Props) {
         }).format(creaatedAt);
 
         return (
-          <Li
-            key={id}
-            className="w-full rounded-md mt-4 p-4 shadow-sm hover:shadow-md transition-shadow duration-150 ease-in-out"
-          >
-            <Wrapper className="flex justify-between">
-              <div>
-                <div>{date}</div>
-                <div>{productName}</div>
-                <div>{productDescription}</div>
-              </div>
-              <div className="w-36 h-36 overflow-hidden rounded-md">
-                <img src={thumbnailUrl} alt="대표 이미지" className="w-full h-full" />
-              </div>
-            </Wrapper>
-            <Wrapper className="flex gap-x-2">
-              <Button type="button">수정</Button>
-              <Button type="button" onClick={() => onClickDeleteButton(id, sellerId)}>
-                삭제
-              </Button>
+          <Li key={id} className="w-full before:h-[2px] before:content-[''] before:block before:bg-slate-100">
+            <Wrapper className="w-full rounded-md p-4">
+              <Wrapper className="flex justify-between">
+                <div>
+                  <div>{date}</div>
+                  <div>
+                    <strong>{productName}</strong>
+                  </div>
+                  <div>{productDescription}</div>
+                </div>
+                <div className="w-36 h-36 overflow-hidden rounded-md">
+                  <img src={thumbnail} alt="대표 이미지" className="object-cover w-full h-full" />
+                </div>
+              </Wrapper>
+              <Wrapper className="flex gap-x-2">
+                <Button type="button" onClick={() => onClickEditButton(id)}>
+                  수정
+                </Button>
+                <Button type="button" onClick={() => onClickDeleteButton(id, sellerId)}>
+                  삭제
+                </Button>
+              </Wrapper>
             </Wrapper>
           </Li>
         );
